refactor: Extract picture rendering related functions to a new file
The functions related to the rendering of the picture have been moved from
viewUpdater to the new file pictureViewUpdater.
